Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #2:

This patch reduces the compiler warning count on Win64 by an additional
550, as reported by VS2013.

+++ client/mysqlbinlog.cc
@@ -1731,7 +1731,7 @@ static my_time_t convert_str_to_timestamp(const char* str)
   long dummy_my_timezone;
   my_bool dummy_in_dst_time_gap;
   /* We require a total specification (date AND time) */
-  if (str_to_datetime(str, (uint) strlen(str), &l_time, 0, &status) ||
+  if (str_to_datetime(str, strlen(str), &l_time, 0, &status) ||
       l_time.time_type != MYSQL_TIMESTAMP_DATETIME || status.warnings)
   {
     error("Incorrect date and time argument: %s", str);
